Added Freight.

+++ api/src/main/java/com/jb/api/Item.java
@@ -8,11 +8,11 @@ import lombok.RequiredArgsConstructor;
 public class Item {
     private final String id;
     private final String description;
-    private final double price;
+    private final Double price;
     private final int width;
     private final int height;
     private final int length;
-    private final int weight;
+    private final Double weight;
 
     public Double getVolume() {
         return (this.width/100.0) * (this.height/100.0) * (this.length/100.0);
